refactor(core,llm): deduplicate resolve_upload_bytes and ToolSpec schema normalization
Extract resolve_upload_bytes into assistant-core::upload so Slack and
Mattermost interfaces share a single implementation instead of
maintaining identical copies.

Add ToolSpec::normalized_params_schema() in assistant-llm so all four
LLM providers (Ollama, Anthropic, OpenAI, Moonshot) call one method
instead of repeating the same schema normalization branch.
