+++ crates/core/Cargo.toml
@@ -15,6 +15,7 @@ chrono = { workspace = true }
 tracing = { workspace = true }
 toml = { workspace = true }
 dirs = { workspace = true }
+base64 = { workspace = true }
 
 [dev-dependencies]
 tempfile = "3"
